+++ plugins/modules/appfwprofile_safeobject_binding.py
@@ -2,7 +2,7 @@
 
 # -*- coding: utf-8 -*-
 
-# Copyright (c) 2023 Cloud Software Group, Inc.
+# Copyright (c) 2025 Cloud Software Group, Inc.
 # MIT License (see LICENSE or https://opensource.org/licenses/MIT)
 
 from __future__ import absolute_import, division, print_function
@@ -76,7 +76,7 @@ options:
     description:
       - Is the rule auto deployed by dynamic profile ?
   maxmatchlength:
-    type: float
+    type: int
     description:
       - Maximum match length for a Safe Object expression.
   name:
