+++ plugins/modules/cmpglobal_cmppolicy_binding.py
@@ -2,7 +2,7 @@
 
 # -*- coding: utf-8 -*-
 
-# Copyright (c) 2023 Cloud Software Group, Inc.
+# Copyright (c) 2025 Cloud Software Group, Inc.
 # MIT License (see LICENSE or https://opensource.org/licenses/MIT)
 
 from __future__ import absolute_import, division, print_function
@@ -103,7 +103,7 @@ options:
     description:
       - The name of the globally bound HTTP compression policy.
   priority:
-    type: float
+    type: int
     description:
       - Positive integer specifying the priority of the policy. The lower the number,
         the higher the priority. By default, polices within a label are evaluated
