+++ plugins/modules/feoparameter.py
@@ -2,7 +2,7 @@
 
 # -*- coding: utf-8 -*-
 
-# Copyright (c) 2023 Cloud Software Group, Inc.
+# Copyright (c) 2025 Cloud Software Group, Inc.
 # MIT License (see LICENSE or https://opensource.org/licenses/MIT)
 
 from __future__ import absolute_import, division, print_function
@@ -39,21 +39,21 @@ options:
       - When C(unset), the resource will be unset on the NetScaler ADC node.
     type: str
   cssinlinethressize:
-    type: float
+    type: int
     description:
       - Threshold value of the file size (in bytes) for converting external CSS files
         to inline CSS files.
   imginlinethressize:
-    type: float
+    type: int
     description:
       - Maximum file size of an image (in bytes), for coverting linked images to inline
         images.
   jpegqualitypercent:
-    type: float
+    type: int
     description:
       - 'The percentage value of a JPEG image quality to be reduced. Range: 0 - 100'
   jsinlinethressize:
-    type: float
+    type: int
     description:
       - Threshold value of the file size (in bytes), for converting external JavaScript
         files to inline JavaScript files.
